Fix git rev-list count
Logic is based on git rev-list to compute the count and set it as a
maximum, such as going through commits one by one stops at the initial
commit.
Git rev-list count shoud not take merges into consideration,
otherwise it is not in sync with 'HEAD~<number>', it is possible to go
over the initial commit, thus reaching a negative scenario.

Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>